refactor: common > message 패키지 이름 변경
