+++ src/main/java/com/programmers/vouchermanagement/message/VoucherMessage.java
@@ -1,4 +1,4 @@
-package com.programmers.vouchermanagement.common;
+package com.programmers.vouchermanagement.message;
 
 public enum VoucherMessage {
     VOUCHER_COMMAND_LIST_MESSAGE("""
