+++ src/main/java/com/programmers/vouchermanagement/message/WalletMessage.java
@@ -1,4 +1,4 @@
-package com.programmers.vouchermanagement.common;
+package com.programmers.vouchermanagement.message;
 
 public enum WalletMessage {
     WALLET_COMMAND_LIST_MESSAGE("""
